refactor(@angular-ru/ng-table-builder): disassemble UtilsService
