--- packages/ng-table-builder/src/services/utils/utils.interface.ts
@@ -1,5 +1,0 @@
-import { PlainObject } from '@angular-ru/common/typings';
-
-export interface UtilsInterface {
-    mergeDeep(...objects: PlainObject[]): PlainObject;
-}
